+++ src/stores/AlertStore.ts
@@ -1,5 +1,7 @@
 import {action, observable, makeObservable} from "mobx";
 import React from "react";
+import {IconName} from "@blueprintjs/icons";
+import {MaybeElement} from "@blueprintjs/core";
 import {Deferred} from "services";
 
 export enum AlertType {
@@ -20,15 +22,17 @@ export class AlertStore {
 
     @observable alertVisible: boolean;
     @observable alertText: string | React.ReactNode;
-    @observable alertIcon: any;
+    @observable alertIcon: IconName | MaybeElement;
     @observable alertType: AlertType;
     @observable interactiveAlertText: string | React.ReactNode;
+    @observable showDashboardLink: boolean;
     private interactionPromise: Deferred<boolean>;
 
-    @action showAlert = (text: string | React.ReactNode, icon?: any) => {
+    @action showAlert = (text: string | React.ReactNode, icon?: IconName | MaybeElement, showDashboard = false) => {
         this.alertText = text;
         this.alertIcon = icon;
         this.alertType = AlertType.Info;
+        this.showDashboardLink = showDashboard;
         this.alertVisible = true;
     };
 
@@ -36,20 +40,22 @@ export class AlertStore {
         this.alertVisible = false;
     };
 
-    @action showInteractiveAlert = (text: string | React.ReactNode, icon?: any) => {
+    @action showInteractiveAlert = (text: string | React.ReactNode, icon?: IconName | MaybeElement, showDashboard = false) => {
         this.interactiveAlertText = text;
         this.alertIcon = icon;
         this.alertType = AlertType.Interactive;
         this.alertVisible = true;
+        this.showDashboardLink = showDashboard;
         this.interactionPromise = new Deferred<boolean>();
         return this.interactionPromise.promise;
     };
 
-    @action showRetryAlert = (text: string | React.ReactNode, icon?: any) => {
+    @action showRetryAlert = (text: string | React.ReactNode, icon?: IconName | MaybeElement, showDashboard = false) => {
         this.interactiveAlertText = text;
         this.alertIcon = icon;
         this.alertType = AlertType.Retry;
         this.alertVisible = true;
+        this.showDashboardLink = showDashboard;
         this.interactionPromise = new Deferred<boolean>();
         return this.interactionPromise.promise;
     };
@@ -65,5 +71,6 @@ export class AlertStore {
     private constructor() {
         makeObservable(this);
         this.alertVisible = false;
+        this.showDashboardLink = false;
     }
 }