+++ src/stores/HelpStore.ts
@@ -4,6 +4,7 @@ import {Position} from "@blueprintjs/core";
 export enum HelpType {
     // Dialogs
     CONTOUR = "Contour Configuration",
+    DISTANCE_MEASUREMENT = "Distance Measurement",
     PREFERENCES = "Preferences",
     FILE_BROWSER = "File Browser",
     FILE_INFO = "File Header",
@@ -12,6 +13,7 @@ export enum HelpType {
     STOKES = "Stokes Hypercube Dialog",
     ONLINE_CATALOG_QUERY = "Online Catalog Query",
     VECTOR_OVERLAY = "Vector Overlay Configuration",
+    IMAGE_FITTING = "Image Fitting",
 
     // Widgets
     ANIMATOR = "Animator",
@@ -51,7 +53,8 @@ export enum HelpType {
     CATALOG_SETTINGS_SIZE = "Catalog Size Settings",
     CATALOG_SETTINGS_ORIENTATION = "Catalog Orientation Settings",
     SPECTRAL_LINE_QUERY = "Spectral Line Query",
-    PV_GENERATOR = "PV Generator"
+    PV_GENERATOR = "PV Generator",
+    CURSOR_INFO = "Cursor Information"
 }
 
 export class HelpStore {